# 🎯 Fix Documentation Accuracy & Remove Artificial Performance Constraints